+++ changelog.md
@@ -1,5 +1,9 @@
 # Stellaris 2.0.1 for MC 26.1.2
 
+## Additions
+- Creative rocket
+- Creative space suit helmet
+
 ## Changes
 - Blender recipe
 - Improve rocket rotation on launch pad
@@ -7,3 +11,4 @@
 ## Fixes
 - Fix every gravity manipulator having the same animation
 - Fix titanium axe, hoe and shovel
+- Error in the console when trying to fill a space suit chestplate with no module